requirements.txt
dnspython==2.8.0
pymongo==4.14.1
python-dotenv==1.1.1